코드 피드백 수정 및 모듈화
