+++ src/js/source/source.js
@@ -1,0 +1,6 @@
+const digit = document.querySelectorAll('.digits > .digit');
+const total = document.querySelector('#total');
+const modifier = document.querySelector('.modifiers > .modifier');
+const operation = document.querySelectorAll('.operations > .operation');
+
+export {digit,total,modifier,operation}
